ai_pwebsite_project/backend/requirements.txt
flask
flask-cors
torch
transformers
sentence-transformers
faiss-cpu
bitsandbytes
numpy
huggingface_hub
accelerate
python-dotenv
